refactor(character position): Refactored character position calculations in the showcase example to...
# Project State

## Current Focus
Refactored character position calculations in the showcase example to improve type safety and clarity.

## Context
The showcase example was refactoring character position calculations to properly handle array indexing and avoid potential overflow issues.

## Completed
- [x] Removed unnecessary `as u32` casts in position calculations
- [x] Simplified boundary checks by using direct integer operations

## In Progress
- [x] No active work in progress for this change

## Blockers
- None identified

## Next Steps
1. Verify the refactored code maintains the same visual output
2. Consider additional optimizations for the position calculations
---
category: refactor
scope: character position
---

+++ tools/tui/dracon-terminal-engine/examples/showcase.rs
@@ -491,8 +491,8 @@ impl Widget for Showcase {
                 if bot_idx < p.cells.len() { p.cells[bot_idx].char = '─'; p.cells[bot_idx].fg = Color::Rgb(0, 200, 150); }
             }
             for y in 0..mh {
-                let l_idx = (my + y) * area.width + mx as u32;
-                let r_idx = l_idx + mw as u32 - 1;
+                let l_idx = (my + y) * area.width + mx;
+                let r_idx = l_idx + mw - 1;
                 if (l_idx as usize) < p.cells.len() { p.cells[l_idx as usize].char = '│'; p.cells[l_idx as usize].fg = Color::Rgb(0, 200, 150); }
                 if (r_idx as usize) < p.cells.len() { p.cells[r_idx as usize].char = '│'; p.cells[r_idx as usize].fg = Color::Rgb(0, 200, 150); }
             }
